Moved current thread thread local to DeterministicRunnerImpl
